# storage: skip MVCCValueHeader parsing in some cases