--- a/pkg/storage/mvcc_value_test.go
+++ b/pkg/storage/mvcc_value_test.go
@@ -400,7 +400,7 @@ func BenchmarkDecodeMVCCValue(b *testing.B) {
 							var ok bool
 							res, ok, err = tryDecodeSimpleMVCCValue(buf)
 							if !ok && err == nil {
-								res, err = decodeExtendedMVCCValue(buf)
+								res, err = decodeExtendedMVCCValue(buf, true)
 							}
 						} else {
 							res, err = DecodeMVCCValue(buf)