Many callers to the functions in pkg/storage do not care about the contents of the MVCCValueHeader. Previously, we would always parse this header anyway. Previously, the impact of this was likely small since most values were unlikely to include an MVCCValueHeader.

However, now IMPORT and LDR both write headers into a large number of rows.

Here, I've added the ability for the caller of decodeExtendedMVCCValue to indicate whether or not they require the value header to be parsed.

In pebbleMVCCScanner we've gone a little further and added a new method decodeValueMVCCValueIgnoringHeader that allows the scanner to avoid the non-inlinable decodeExtnededMVCCValue.

In the added benchmark case, this has been observed to produce a 7-10% improvement:

```
                                                                              │ before.txt  │              after.txt              │
                                                                              │   sec/op    │   sec/op     vs base                │
MVCCScan_Pebble/rows=1000/versions=2/valueSize=64/numRangeKeys=0/headers=true   403.5µ ± 1%   362.7µ ± 1%  -10.13% (p=0.000 n=20)

                                                                              │  before.txt  │              after.txt               │
                                                                              │     B/s      │     B/s       vs base                │
MVCCScan_Pebble/rows=1000/versions=2/valueSize=64/numRangeKeys=0/headers=true   151.2Mi ± 1%   168.3Mi ± 1%  +11.27% (p=0.000 n=20)
```

No larger scale testing of this has been done.

If the code looks a bit tortured to your eye, I agree. I've attempted to make sure that the same amount of inlining was happening. This is a bit finicky since many of the functions involved are very close to the default complexity limit for inlining functions.

Informs #131359

Epic: none
Release note: None